Resolved conflicts:
	include/cinder/Cinder.h
	include/cinder/audio/msw/ContextWasapi.h
	include/cinder/audio/msw/DeviceManagerWasapi.h
	vc2015_winrt/cinder.vcxproj

+++ blocks/Cairo/include/cinder/cairo/Cairo.h
@@ -788,7 +788,7 @@ namespace cinder {
 class SurfaceConstraintsCairo : public cinder::SurfaceConstraints {
  public:
 	virtual SurfaceChannelOrder getChannelOrder( bool alpha ) const;
-	virtual int32_t				getRowBytes( int requestedWidth, const SurfaceChannelOrder &sco, int elementSize ) const;
+	virtual ptrdiff_t			getRowBytes( int requestedWidth, const SurfaceChannelOrder &sco, int elementSize ) const;
 };
 
 } // namespace cinder
